% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesianPosteriorProbability.R
\name{bayesianPosteriorProbability}
\alias{bayesianPosteriorProbability}
\alias{bayesianPosteriorProbabilityPlot}
\title{Visualize the distribution of Bayesian posterior probabilities in ggplot}
\usage{
bayesianPosteriorProbability(predictedOutcomes = c("HAND", "NonHAND"),
  priorProbHAND = 0.33, confMat = NULL, sensitivity = NULL,
  specificity = NULL)

bayesianPosteriorProbabilityPlot(predictedOutcomes = c("HAND", "NonHAND"),
  confMat = NULL, sensitivity = NULL, specificity = NULL)
}
\arguments{
\item{predictedOutcomes}{The prediction result. Either "HAND", "NonHAND", or both.}

\item{priorProbHAND}{Prior probability of HAND. It is the user's responsibility to provide a rational estimate. The default is set to be 0.33, which is derived from The Multicenter AIDS Cohort Study [PMID: 26718568].}

\item{confMat}{The caret confusion matrix of the stacked classifier.}

\item{sensitivity}{(Optional) The sensitivity of the stacked classifier.}

\item{specificity}{(Optional) The specificity of the stacked classifier.}
}
