% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataImportAndCombine.R
\name{dataImportAndCombine}
\alias{dataImportAndCombine}
\title{Load & combine a sequence dataset and a metadata into a single dataframe}
\usage{
dataImportAndCombine(fileName.Alignment.FASTA, fileName.HIVNeuroMetadata.CSV,
  fileName.SampleTissueCategoryDesignSheet.CSV, colName.SequenceID,
  colName.Sample.Tissue, colName.Clinical.Status = NULL)
}
\arguments{
\item{fileName.Alignment.FASTA}{A file name of the HIV env sequence alignment in a FASTA format.}

\item{fileName.HIVNeuroMetadata.CSV}{A file name of the metadata in a CSV format.}

\item{fileName.SampleTissueCategoryDesignSheet.CSV}{A file name of the tissue category design sheet in a CSV format.}

\item{colName.SequenceID}{The column name representing sequene ID in the metadata.}

\item{colName.Sample.Tissue}{The column name representing sample source tissues in the metadata.}

\item{colName.Clinical.Status}{(Optional) The column name representing neurological outcomess in the metadata.}
}
