% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggPie.R
\name{ggPie}
\alias{ggPie}
\title{Pie chart in ggplot.}
\usage{
ggPie(labelSet, colorSet)
}
\arguments{
\item{labelSet}{A vector of categorical variables.}

\item{colorSet}{The color set to be used.}
}
\author{
stack overflow
}
