% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggRF.R
\name{ggRF}
\alias{ggRF}
\title{Visualize Random Forest model in ggplot}
\usage{
ggRF(
  caret_rf_model,
  num_nodes = "median",
  brewerPalName = "Set1",
  lastColorGrey = F
)
}
\arguments{
\item{caret_rf_model}{The final model of the random forest model from caret}

\item{num_nodes}{The number of nodes for visualization. Either "smallest", "largest", "median", or an integer.}

\item{brewerPalName}{The color palette name in the Brewer palettes.}

\item{lastColorGrey}{Logical. Whether the last category should be colored as grey.}
}
\description{
Visualize Random Forest model in ggplot
}
