% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gVisChart.R
\name{gvSankey}
\alias{gvSankey}
\title{Comparison of reference and prediction in sankey plot using googleVis API.}
\usage{
gvSankey(reference, prediction, colors_link)
}
\arguments{
\item{reference}{A set of reference labels.}

\item{prediction}{A set of predicted labels.}

\item{colors_link}{A set of colors.}
}
