% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machineLearning.Comparison.R
\name{machineLearning.Comparison.default}
\alias{machineLearning.Comparison.default}
\alias{machineLearning.Comparison}
\alias{machineLearning.Comparison.Bootstrap}
\title{Compare multiple machine learning algorithms.}
\usage{
machineLearning.Comparison.default(
  trainingData,
  testingData,
  outcomeLabelName,
  algorithmList,
  algorithmLabelList,
  trainControlOptions,
  stackingAlgorithm = "xgbTree",
  tuneL = 3
)

machineLearning.Comparison(
  ml.data,
  ml.metadata,
  holdoutTrainingRatio = 0.8,
  outcomeLabelName,
  patientIDLabelName,
  algorithmList,
  algorithmLabelList,
  trainControlOptions,
  stackingAlgorithm = "xgbTree",
  tuneL = 3,
  outputFileName.TXT = NULL,
  seed = 123456789
)

machineLearning.Comparison.Bootstrap(..., seedList = 1:10)
}
\arguments{
\item{trainingData}{A training dataset.}

\item{testingData}{A testing dataset.}

\item{outcomeLabelName}{A column name for neurological outcomes.}

\item{algorithmList}{Machine learning algorithms to be compared.}

\item{algorithmLabelList}{User-defined labels for the machine learning algorithms to be compared.}

\item{trainControlOptions}{A trainControl object for the train function in caret.}

\item{stackingAlgorithm}{An algorithm used for model stacking.}

\item{tuneL}{An integer denoting the amount of granularity in the grid search of tuning parameters in the train function.}

\item{ml.data}{A machine-learning-ready dataframe returned by the machineLearning.DataFormat function.}

\item{ml.metadata}{A metadata.}

\item{holdoutTrainingRatio}{A ratio of data to be used for model training.}

\item{patientIDLabelName}{A column name for patient IDs.}

\item{outputFileName.TXT}{(Optional) Print and save the comparison results in a TXT format.}

\item{seed}{A random seed.}

\item{...}{Parameters for the machineLearning.Comparison function.}

\item{seedList}{A set of random seeds.}
}
\description{
Compare multiple machine learning algorithms.
}
