% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machineLearning.DataFormatting.R
\name{machineLearning.DataFormat}
\alias{machineLearning.DataFormat}
\title{Convert the sequence-derived numerical matrix into a machine-learning-ready dataframe.}
\usage{
machineLearning.DataFormat(ml.matrix, ml.metadata, colNames.from,
  colNames.to = colNames.from, outcomeLabelName = NULL, sourceLabelName,
  pp = "internal", pp.by.sequence = NULL, pp.by.patient = NULL)
}
\arguments{
\item{ml.matrix}{A numerical matrix returned by the machineLearning.Matrix function.}

\item{ml.metadata}{A metadata.}

\item{colNames.from}{Names of the columns in the metadata to be incorporated in the subsequent analysis.}

\item{colNames.to}{(Optional) Column names of the included metadata. Default values are the same as colNames.from.}

\item{outcomeLabelName}{A column name representing the neurological outcomes.}

\item{sourceLabelName}{A column name representing the sample sources.}

\item{pp}{An indicator for preprocessing. If set "internal", variables are first screened using the zero-variance and near-zero-variance methods implemented in the caret package, and then further screened using the correlation method implemented in the caret package. If set "external", the matrix is pre-processed using externally provided functions. Omit preprocessing by setting "none".}

\item{pp.by.sequence}{(Optional) Users can provide their own pre-processing function (per sequence).}

\item{pp.by.patient}{(Optional) Users can provide their own pre-processing function (per patient).}
}
