% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machineLearning.DataFormatting.R
\name{aaseqToBoolian}
\alias{aaseqToBoolian}
\alias{aaseqToBoolian.batch}
\alias{aaseqToAAIndex}
\alias{aaseqToAAIndex.batch}
\alias{machineLearning.Matrix}
\title{Convert an amino acid sequence to a numerical matrix.}
\usage{
aaseqToBoolian(sequence)

aaseqToBoolian.batch(sequences, gapPositions = c(125))

aaseqToAAIndex(sequence, aaIndexType = "all", normalized = T)

aaseqToAAIndex.batch(sequences, gapPositions = c(125), aaIndexType = "all",
  normalized = T)

machineLearning.Matrix(sequences, gapPositions = c(125),
  analysisType = "AAIndex", aaIndexType = "physicochemical",
  normalized = T)
}
\arguments{
\item{sequence}{An amino acid sequence of HIV C2V3C3 region.}

\item{sequences}{A set of amino acid sequences.}

\item{gapPositions}{Alignment gap positions.}

\item{aaIndexType}{An indicator for the type of the set of AAIndex to be used. Currently, "all" and "physicochemical" are implemented. If set to be "all", a total of 544 AAIndices included in the BioSeqClass package are used. If set to be "physicochemical", pre-selected 78 AAIndices are used. See the original paper for further details.}

\item{normalized}{Logical. Whether AAIndex values are rescaled to [0,1] range.}

\item{analysisType}{An indicator for the type of  analysis. Currently, "boolian" and "AAIndex" are implemented.}
}
