% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machineLearning.Prediction.R
\name{machineLearning.Stacking.Predict}
\alias{machineLearning.Stacking.Predict}
\alias{machineLearning.Stacking.Predict_From_PreprocessedDataset}
\alias{machineLearning.Stacking.Predict_From_Metadataset}
\title{Predict HAND status using pre-trained classifiers.}
\usage{
machineLearning.Stacking.Predict(X, ...)

machineLearning.Stacking.Predict_From_PreprocessedDataset(
  ml.data.predict,
  modelList,
  model.stack
)

machineLearning.Stacking.Predict_From_Metadataset(
  sequences,
  gapPositions = c(125),
  analysisType = "AAIndex",
  aaIndexType = "physicochemical",
  normalized = T,
  ml.metadata,
  colNames.from,
  colNames.to,
  outcomeLabelName,
  sourceLabelName,
  pp = "external",
  pp.by.sequence,
  pp.by.patient,
  modelList,
  model.stack
)
}
\arguments{
\item{X}{Either a dataframe or a set of sequences.}

\item{...}{Parameters passed to an appropriate working function.}

\item{ml.data.predict}{A machine-learning-ready dataframe,}

\item{modelList}{A list of pre-trained models.}

\item{model.stack}{A pre-trained stacking model.}

\item{sequences}{A set of amino acid sequences of HIV C2V3C3 region.}

\item{gapPositions}{Alignment gap positions.}

\item{analysisType}{An indicator for the type of  analysis. Currently, "boolian" and "AAIndex" are implemented.}

\item{aaIndexType}{Currently, "all" and "physicochemical" are implemented. If set to be "all", a total of 544 AAIndices included in the BioSeqClass package are used. If set to be "physicochemical", pre-selected 78 AAIndices are used. See the original paper for further details.}

\item{normalized}{Logical. Whether AAIndex values are rescaled to [0,1] range.}

\item{ml.metadata}{A metadata.}

\item{colNames.from}{Names of the columns in the metadata to be incorporated in the subsequent analysis.}

\item{colNames.to}{(Optional) Column names of the included metadata. Default values are the same as colNames.from.}

\item{outcomeLabelName}{A column name representing the neurological outcomes.}

\item{sourceLabelName}{A column name representing the sample sources.}

\item{pp}{An indicator for preprocessing. If set "internal", variables are first screened using the zero-variance and near-zero-variance methods implemented in the caret package, and then further screened using the correlation method implemented in the caret package. If set "external", the matrix is pre-processed using externally provided functions. Omit preprocessing by setting "none".}

\item{pp.by.sequence}{(Optional) Users can provide their own pre-processing function (per sequence).}

\item{pp.by.patient}{(Optional) Users can provide their own pre-processing function (per patient).}
}
\description{
Predict HAND status using pre-trained classifiers.
}
