% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machineLearning.FeatureImportance.R
\name{machineLearning.ResidueAAIndexDualPlot}
\alias{machineLearning.ResidueAAIndexDualPlot}
\title{Residue frequencies and AAIndex values at specified important positions.}
\usage{
machineLearning.ResidueAAIndexDualPlot(
  featureSet,
  sequences,
  seqWeights = NULL,
  outcomes,
  xLabel = NULL,
  yLabel1 = NULL,
  yLabel2 = NULL,
  legendLabel = NULL,
  colorSet
)
}
\arguments{
\item{featureSet}{An important feature set.}

\item{sequences}{Sequences to be analyzed.}

\item{seqWeights}{Weights per sequence. Patient-to-patient differences in sequencing depth should be normalized. Can be provided either as fractions (<1) or sequence counts (>1).}

\item{outcomes}{A character vector containing neurological outcomes.}

\item{xLabel}{An x-axis label.}

\item{yLabel1}{A y1-axis label [Residue frequency].}

\item{yLabel2}{A y2-axis label [AAIndex value].}

\item{legendLabel}{A legend label.}

\item{colorSet}{The color set to be used.}
}
\description{
Residue frequencies and AAIndex values at specified important positions.
}
