% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machineLearning.StepwiseFeatureReduction.R
\name{machineLearning.StepwiseFeatureReduction}
\alias{machineLearning.StepwiseFeatureReduction}
\title{Stepwise feature reduction.}
\usage{
machineLearning.StepwiseFeatureReduction(featureSet, ml.data, outcomeLabelName,
  ...)
}
\arguments{
\item{featureSet}{A set of features as a starting point. Can be either a character vector or a list.}

\item{ml.data}{A machine-learning-ready dataframe returned by the machineLearning.DataFormat function.}

\item{outcomeLabelName}{A column name for neurological outcomes.}

\item{...}{Parameters for the machineLearning.Comparison.Bootstrap function.}
}
