% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machineLearning.FeatureImportance.R
\name{machineLearning.FeatureImportanceAnalysis}
\alias{machineLearning.FeatureImportanceAnalysis}
\title{Screen important features.}
\usage{
machineLearning.FeatureImportanceAnalysis(ml.data, outcomeLabelName, modelList,
  topN = 20, pThreshold = 0.05, removedVars = c(),
  outputFileName.DOCX = "FeatureImportance.docx", xLabel = NULL,
  yLabel = NULL, colorSet)
}
\arguments{
\item{ml.data}{A machine-learning-ready dataframe returned by the machineLearning.DataFormat function.}

\item{outcomeLabelName}{A column name for neurological outcomes.}

\item{modelList}{A list of caret models.}

\item{topN}{An integer indicating how many the most important features are retained.}

\item{pThreshold}{A threshold of significance. Features with significantly distinct value distributions between neurological outcomes are considered important.}

\item{removedVars}{(Optional) Feature names that should be removed manually.}

\item{outputFileName.DOCX}{A file name of a summary table in a Word format.}

\item{xLabel}{An x-axis label for the violin plot.}

\item{yLabel}{A y-axis label for the violin plot.}

\item{colorSet}{The color set to be used.}
}
